# docs/openapi.yaml
openapi: 3.0.0
info:
  title: Email Validator API
  description: |
    A robust HTTP API for validating email addresses through various methods including syntax checking, 
    DNS verification, and SMTP validation.
  version: 1.0.0
servers:
  - url: https://api.emailvalidator.com/v1
    description: Production server
  - url: http://localhost:8000/v1
    description: Local development server
security:
  - ApiKeyAuth: []
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
...
    BatchValidationRequest:
      type: object
      required:
        - emails
      properties:
        emails:
          type: array
          items:
            type: string
            format: email
          maxItems: 100
          example: ["user1@example.com", "user2@example.com"]
    ValidationResponse:
      type: object
      properties:
        is_valid:
          type: boolean
        checks:
          type: object
          properties:
            format:
              type: boolean
            dns:
              type: boolean
            smtp:
              type: boolean
        error:
          type: string
          nullable: true
    BatchValidationResponse:
      type: object
      properties:
        results:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/ValidationResponse'
...
paths:
...
  /validate/batch:
    post:
      summary: Validate multiple email addresses
      description: Validates multiple email addresses in a single request
      operationId: validateEmailBatch
      tags:
        - Validation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BatchValidationRequest'
      responses:
        '200':
          description: Batch validation results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BatchValidationResponse'
        '400':
          description: Invalid request format
        '429':
          description: Too many requests
        '500':
          description: Internal server error
...